Refactor Torrent component
