+++ src/components/App/styles/index.css
@@ -1,4 +1,20 @@
 :global body {
   margin: 0px;
   padding: 0px;
+
+  font: 62.5% "lucida grande", Tahoma, Verdana, Arial, Helvetica, sans-serif;
+}
+
+.torrentList {
+  width: 100%;
+  margin: 0;
+  padding: 0;
+  text-align: left;
+  cursor: pointer;
+}
+
+.torrentRow {
+  border-bottom: 1px solid #ccc;
+  padding: 5px 10px;
+  background-color: white;
 }
