--- src/components/Torrent/PeerDetails/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { getPeerDetails } from './services/formatters';
-
-function PeerDetails({ torrent }) {
-  return (
-    <p>{getPeerDetails(torrent)}</p>
-  );
-}
-
-export default PeerDetails;
