--- src/components/Torrent/ProgressBar.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function ProgressBar({ torrent }) {
-  return (
-    <progress max="100" value="80"></progress>
-  );
-}
-
-export default ProgressBar;
