+++ src/components/Torrent/ProgressBar/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import CSSModules from 'react-css-modules';
+
+import styles from './styles';
+
+function ProgressBar({ torrent }) {
+  return (
+    <div styleName='torrentProgressBar'>
+      <progress max="100" value="80" />
+    </div>
+  );
+}
+
+export default CSSModules(ProgressBar, styles);
