--- src/components/Torrent/ProgressDetails.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function ProgressDetails({ torrent }) {
-  return (
-    <p>Downloading ... ↓ {torrent.rateDownload} ↑ {torrent.rateUpload}</p>
-  );
-}
-
-export default ProgressDetails;
