+++ src/components/Torrent/services/formatters.js
