+++ src/components/Torrent/styles/index.css
@@ -1,5 +1,42 @@
 .torrent {
-  margin: 10px;
-  border: 1px solid grey;
-  padding: 5px;
+  color: #666;
+}
+
+.name {
+  font-size: 1.3em;
+  font-weight: bold;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  color: #222;
+  margin-top: 2px;
+  margin-bottom: 2px;
+}
+
+.peerDetails {
+  clear: left;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.progressBarRow {
+  display: flex;
+}
+
+.progressBarRow div {
+  flex: 1;
+}
+
+.progressBarRow button {
+  width: 14px;
+  height: 14px;
+  margin-left: 8px;
+}
+
+.progressDetails {
+  clear: left;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
 }
